cambios en nuevo espacio y modificar espacio

+++ src/main/kotlin/ar/edu/unsam/pds/controller/dto/UsuarioDTO.kt
@@ -17,7 +17,7 @@ data class UsuarioDTO(
 )
 
 data class UsuarioIdDTO(
-    val id: Int? = null,
+    val id: Long? = null,
 )
 
-fun Usuario.toIdDTO() = UsuarioIdDTO(id!!.toInt())
+fun Usuario.toIdDTO() = UsuarioIdDTO(id!!)
